+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * **breaking** renderer: Make `Sprite::render` public, and `Sprite::new_from_image` takes the image
     by reference instead of by value.
 * riddle: Add rustfmt.toml, setting `hard_tabs = true`. Reformat repo.
+* **breaking** renderer: Remove `SpriteRenderArgs::render`, instead prefer use of `Sprite::render`.
 
 ## 0.2.0